Moved `logger.js` inside paymentservice and resolved `header-check` issue
